\name{GoFKernel-package}
\alias{GoFKernel-package}
\alias{GoFKernel}
\docType{package}
\title{
GoFKernel: Testing Goodness-of-fit with the Kernel Density Estimator
}
\description{
Tests of goodness-of-fit based on kernel smoothing of the data.
}
\details{
\tabular{ll}{
Package: \tab GoFKernel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-22\cr
License: \tab GPL\cr
Depends: \tab R (>=2.15.3), \href{http://stat.ethz.ch/R-manual/R-patched/library/stats/html/00Index.html}{stats}, \href{http://cran.r-project.org/web/packages/KernSmooth/index.html}{KernSmooth} (>=2.23-8)
}
The most important functions in \code{GoFKernel} are \code{\link{dgeometric.test}} and \code{\link{fan.test}}.
}
\author{
Jose M. Pavia

Maintainer: Jose M. Pavia <Jose.M.Pavia@uv.es>

}
\references{
Fan, Y (1994) "Testing the goodness-of-fit of a parametric density function by kernel method",
\emph{Econometric Theory}, 10, 316-356.\cr
Li, O. and Racine, J.F. (2007) \emph{Nonparametric Econometrics}, Princeton University Press, New Jersey.\cr
Pavia, JM (2013) "Testing Goodness-of-fit with the Kernel Density Estimator: GoFKernel".\cr
}
\keyword{ package }
