\name{GoFKernel-package}
\alias{GoFKernel-package}
\alias{GoFKernel}
\docType{package}
\title{
Testing Goodness-of-fit with the Kernel Density Estimator
}
\description{
Tests of goodness-of-fit based on kernel smoothing of the data.
}
\details{
\tabular{ll}{
Package: \tab GoFKernel\cr
Depends: \tab R (>=2.17.3), \code{stats}, \code{KernSmooth} (>=2.23-8)\cr
Type: \tab Package\cr
Version: \tab 2.0-5\cr
Date: \tab 2014-12-03\cr
License: \tab GPL\cr
}
The most important functions in \code{GoFKernel} are \code{\link{dgeometric.test}} and \code{\link{fan.test}}.
}
\author{
Jose M. Pavia

Maintainer: Jose M. Pavia <Jose.M.Pavia@uv.es>

}
\references{
Fan, Y (1994) "Testing the goodness-of-fit of a parametric density function by kernel method", Econometric Theory, 10, 316-356.

Li, O. and Racine, J.F. (2007) "Nonparametric Econometrics", Princeton University Press, New Jersey.

Pavia, JM (2014) "Testing Goodness-of-fit with the Kernel Density Estimator: GoFKernel".
}

\keyword{ package }
